Fix migrations to use revert and run properly
